creation feature/profile-view

+++ backend/internal/repository/user_repository.go
@@ -9,7 +9,7 @@ import (
 
 type UserRepository interface {
 	Create(ctx context.Context, user *model.User) error
-	GetByID(ctx context.Context, id int64) (*model.User, error)
+	GetByID(ctx context.Context, id int) (*model.User, error)
 	GetByEmail(ctx context.Context, email string) (*model.User, error)
 }
 
@@ -44,7 +44,7 @@ func (r *userRepository) Create(ctx context.Context, user *model.User) error {
 	return nil
 }
 
-func (r *userRepository) GetByID(ctx context.Context, id int64) (*model.User, error) {
+func (r *userRepository) GetByID(ctx context.Context, id int) (*model.User, error) {
 	query := `
 		SELECT id, username, email, password_hash, created_at
 		FROM users
